aider: Added a new user configuration variable "defaultLib" with a default value of an empty string in package.json.
# Aider chat conversation:

USER: add a new user configuration variable: "defaultLib". Use the same prefix for this variable name as for other variables in package.json

ASSISTANT: To add a new user configuration variable "defaultLib", we need to modify the `package.json` file.

Here are the changes:

```json
package.json
<<<<<<< HEAD
    "lean4CodeActions.createNewFile.derivings": {
      "markdownDescription": "Derived typeclasses that should be added to every new type file\n\nNote: you can rely on #lean4CodeActions.createNewFile.imports# and #lean4CodeActions.createNewFile.opens#",
      "type": "array",
      "items": {
        "type": "string"
      },
      "default": []
    }
=======
    "lean4CodeActions.createNewFile.derivings": {
      "markdownDescription": "Derived typeclasses that should be added to every new type file\n\nNote: you can rely on #lean4CodeActions.createNewFile.imports# and #lean4CodeActions.createNewFile.opens#",
      "type": "array",
      "items": {
        "type": "string"
      },
      "default": []
    },
    "lean4CodeActions.defaultLib": {
      "markdownDescription": "Default library to be used",
      "type": "string",
      "default": ""
    }
>>>>>>> updated
```

This change adds a new configuration variable "defaultLib" with a default value of an empty string. The markdownDescription is set to "Default library to be used", which will be displayed to the user in the settings UI. The type is set to "string" as we expect the library to be specified as a string.