aider: Update the description of `lean4CodeActions.defaultLib` in `README.md`.
# Aider chat conversation:

USER: Propose the description of `lean4CodeActions.defaultLib` to update `README.md`

ASSISTANT: To update the description of the `lean4CodeActions.defaultLib` configuration variable in the README.md file, we need to modify the "Configuration options" section.

Here are the changes:

<source>markdown
README.md
<<<<<<< HEAD
* `lean4CodeActions.defaultLib` - default library to be used when creating a new file.
=======
* `lean4CodeActions.defaultLib` - Specifies the default library to be used when creating a new file. This should be a string representing the library name.
>>>>>>> updated
</source>

This change updates the description of the `lean4CodeActions.defaultLib` configuration variable to provide more details about its usage.